requirements.txt
streamlit>=1.34
chromadb>=0.5

langchain>=0.3
langchain-core>=0.3
langchain-openai>=0.2
langchain-chroma>=0.1.2
langchain-text-splitters>=0.3
langchain-community>=0.3

# v1.7: Voyage AI embeddings (voyage-3-large)
# 200M tokens FREE, surpasses voyage-law-2 on legal benchmarks
langchain-voyageai>=0.1

# v1.6: Token-aware chunking
tiktoken>=0.7

# v1.6: Hybrid BM25+Dense retrieval
rank_bm25>=0.2.2

pydantic>=2.0
python-dotenv>=1.0

beautifulsoup4>=4.12
lxml>=5.0
pandas>=2.2

tenacity>=8.2

# MIME validation (libmagic). If libmagic missing, app falls back gracefully.
python-magic>=0.4.27

pytest>=8.0
ruff>=0.6
